[#29] refactor : 클래스 분리
- 리팩토링 과정의 일부로서 클래스 분리

-------------------------
-------------------------
