+++ BE/src/main/java/todo3/codesquad/Controller/TodoController.java
@@ -1,4 +1,4 @@
-package todo3.codesquad;
+package todo3.codesquad.Controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -7,6 +7,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import todo3.codesquad.domain.Response;
+import todo3.codesquad.message.ResponseMessage;
+import todo3.codesquad.domain.Card;
+import todo3.codesquad.domain.CardRepository;
+import todo3.codesquad.domain.Col;
+import todo3.codesquad.domain.ColRepository;
 
 import java.util.ArrayList;
 import java.util.List;
