+++ BE/src/main/java/todo3/codesquad/domain/CardRepository.java
@@ -1,4 +1,4 @@
-package todo3.codesquad;
+package todo3.codesquad.domain;
 
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
