+++ BE/src/main/java/todo3/codesquad/domain/Col.java
@@ -1,4 +1,4 @@
-package todo3.codesquad;
+package todo3.codesquad.domain;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
