+++ BE/src/main/java/todo3/codesquad/domain/Response.java
@@ -1,8 +1,9 @@
-package todo3.codesquad;
+package todo3.codesquad.domain;
 
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import todo3.codesquad.domain.Card;
 
 import java.util.List;
